moved folium plotting to own file
